add function comprassion

+++ uu_imenu_maing01-server/app/api/validation_types/ingredience-types.js
@@ -1,6 +1,6 @@
 const ingredienceCreateDtoInType = shape({
   name: string(255).isRequired(),
-  amount: integer(400).isRequired(),
+  amount: float(400).isRequired(),
   unit: string(400).isRequired(),
   allergen: boolean().isRequired(),
   category: string(255).isRequired(),
@@ -15,7 +15,7 @@ const ingredienceDeleteDtoInType = shape({
 const ingredienceUpdateDtoInType = shape({
   id: id().isRequired(),
   name: string(255),
-  amount: integer(400),
+  amount: float(400),
   unit: string(400),
   allergen: string(100),
   category: string(255),
